third_party/blink/web_tests/external/wpt/editing/run/multitest_6001-7000-expected.txt

This is a testharness.js-based test.
Found 119 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["fontsize","4"],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><font size=\\"4\\">a</font>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["fontsize","4"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"font-size:large\\"><br></span>bar"
[FAIL] [["fontsize","4"],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><font size=\\"4\\">a</font>bar" but got "foo<span style=\\"font-size:large\\"><br>a</span>bar"
[FAIL] [["fontsize","4"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","4"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","4"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<font size=\\"4\\">a</font>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["fontsize","4"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","4"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","4"],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["fontsize","4"],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><font size=\\"4\\">a</font>bar</p>" but got "foo<div><span style=\\"font-size:large\\">a</span>bar</div>"
[FAIL] [["fontsize","4"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","4"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","4"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<font size=\\"4\\">a</font>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["fontsize","4"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["fontsize","4"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["fontsize","4"],["justifycenter",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["fontsize","4"],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<font size=\\"4\\">a</font>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["fontsize","4"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["fontsize","4"],["justifyfull",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["fontsize","4"],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<font size=\\"4\\">a</font>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["fontsize","4"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["fontsize","4"],["justifyleft",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontsize","4"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontsize","4"],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>bar" but got "fooabar"
[FAIL] [["fontsize","4"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontsize","4"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["fontsize","4"],["justifyright",""]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["fontsize","4"],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<font size=\\"4\\">a</font>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["fontsize","4"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("fontsize") after
  assert_equals: Wrong result returned expected "4" but got "3"
[FAIL] [["fontsize","4"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["fontsize","4"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["fontsize","4"],["outdent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font size=\\"4\\">a</font>bar" but got "foo<span style=\\"font-size:large\\">a</span>bar"
[FAIL] [["forecolor","#0000FF"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">a</font>bar" but got "foo<span style=\\"color:rgb(0, 0, 255)\\">a</span>bar"
[FAIL] [["forecolor","#0000FF"],["delete",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["delete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fo<font color=\\"#0000ff\\">a</font>bar" but got "foabar"
[FAIL] [["forecolor","#0000FF"],["delete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["formatblock","<div>"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>foo<font color=\\"#0000ff\\">a</font>bar</div>" but got "<div>fooabar</div>"
[FAIL] [["forecolor","#0000FF"],["formatblock","<div>"],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["forwarddelete",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">a</font>ar" but got "fooaar"
[FAIL] [["forecolor","#0000FF"],["forwarddelete",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["indent",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foobar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">foobar</blockquote>"
[FAIL] [["forecolor","#0000FF"],["indent",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["indent",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<blockquote>foo<font color=\\"#0000ff\\">a</font>bar</blockquote>" but got "<blockquote style=\\"margin:0 0 0 40px; border:none; padding:0px\\">fooabar</blockquote>"
[FAIL] [["forecolor","#0000FF"],["indent",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["inserthorizontalrule",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<hr><font color=\\"#0000ff\\">a</font>bar" but got "foo<hr>abar"
[FAIL] [["forecolor","#0000FF"],["inserthorizontalrule",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["inserthtml","ab<b>c</b>d"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "fooab<b>c</b>d<font color=\\"#0000ff\\">a</font>bar" but got "fooab<b>c</b>dabar"
[FAIL] [["forecolor","#0000FF"],["inserthtml","ab<b>c</b>d"],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertimage","/img/lion.svg"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<img src=\\"/img/lion.svg\\"><font color=\\"#0000ff\\">a</font>bar" but got "foo<img src=\\"/img/lion.svg\\">abar"
[FAIL] [["forecolor","#0000FF"],["insertimage","/img/lion.svg"],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertlinebreak",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br>bar" but got "foo<span style=\\"color:rgb(0, 0, 255)\\"><br></span>bar"
[FAIL] [["forecolor","#0000FF"],["insertlinebreak",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<br><font color=\\"#0000ff\\">a</font>bar" but got "foo<span style=\\"color:rgb(0, 0, 255)\\"><br>a</span>bar"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ol><li>foo<font color=\\"#0000ff\\">a</font>bar</li></ol>" but got "<ol><li>afoobar</li></ol>"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["forecolor","#0000FF"],["insertorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["forecolor","#0000FF"],["insertparagraph",""]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p>bar</p>" but got "foo<div>bar</div>"
[FAIL] [["forecolor","#0000FF"],["insertparagraph",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p>foo</p><p><font color=\\"#0000ff\\">a</font>bar</p>" but got "foo<div><span style=\\"color:rgb(0, 0, 255)\\">a</span>bar</div>"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<ul><li>foo<font color=\\"#0000ff\\">a</font>bar</li></ul>" but got "<ul><li>afoobar</li></ul>"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["forecolor","#0000FF"],["insertunorderedlist",""],["inserttext","a"]] "foo[]bar" queryCommandValue("insertunorderedlist") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:center\\">foo<font color=\\"#0000ff\\">a</font>bar</div>" but got "<div style=\\"text-align:center\\">fooabar</div>"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifycenter",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifycenter") after
  assert_equals: Wrong result returned expected "center" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:justify\\">foo<font color=\\"#0000ff\\">a</font>bar</div>" but got "<div style=\\"text-align:justify\\">fooabar</div>"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifyfull",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyfull") after
  assert_equals: Wrong result returned expected "justify" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<font color=\\"#0000ff\\">a</font>bar" but got "fooabar"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") before
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyleft",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyleft") after
  assert_equals: Wrong result returned expected "left" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyright",""]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifyright",""]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
[FAIL] [["forecolor","#0000FF"],["justifyright",""],["inserttext","a"]] "foo[]bar" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"text-align:right\\">foo<font color=\\"#0000ff\\">a</font>bar</div>" but got "<div style=\\"text-align:right\\">fooabar</div>"
[FAIL] [["forecolor","#0000FF"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("forecolor") after
  assert_equals: Wrong result returned (after color normalization) expected "rgb(0, 0, 255)" but got "rgb(0, 0, 0)"
[FAIL] [["forecolor","#0000FF"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") before
  assert_equals: Wrong result returned expected "left" but got "false"
[FAIL] [["forecolor","#0000FF"],["justifyright",""],["inserttext","a"]] "foo[]bar" queryCommandValue("justifyright") after
  assert_equals: Wrong result returned expected "right" but got "true"
Harness: the test ran to completion.
